requirements.txt
argcomplete==1.10.3
attrs==21.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
blis==0.7.5
boto3==1.20.0
botocore==1.23.0
catalogue==1.0.0
certifi==2021.10.8
cffi==1.15.0
cfn-flip==1.3.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
compressed-rtf==1.0.6
cryptography==35.0.0
cycler==0.11.0
cymem==2.0.6
deep-translator==1.5.4
docopt==0.6.2
docx2txt==0.8
durationpy==0.5
ebcdic==1.1.1
EbookLib==0.17.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
extract-msg==0.28.7
Flask==1.1.2
future==0.18.2
gensim==3.6.0
hjson==3.0.2
idna==3.3
IMAPClient==2.1.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.2.1
kappa==0.6.0
kiwisolver==1.3.2
langcodes==3.2.1
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.4.3
murmurhash==1.0.6
mysql-connector-python==8.0.27
nltk==3.6.5
numpy==1.21.4
olefile==0.46
packaging==21.2
pandas==1.3.4
pathy==0.6.1
pdfminer==20191125
pdfminer.six==20211012
pep517==0.12.0
Pillow==8.4.0
pip-tools==6.4.0
pipreqs==0.4.11
plac==1.1.3
placebo==0.10.0
plotly==4.4.1
preshed==3.0.6
protobuf==3.19.1
pycparser==2.21
pycryptodome==3.11.0
pydantic==1.8.2
pyparsing==2.4.7
pyresparser==1.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
python-slugify==5.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.2
requests==2.26.0
retrying==1.3.3
s3transfer==0.5.0
scikit-learn==1.0.1
scipy==1.7.2
six==1.12.0
sklearn==0.0
smart-open==5.2.1
sortedcontainers==2.4.0
soupsieve==2.3
spacy==2.3.5
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SpeechRecognition==3.8.1
srsly==1.0.5
text-unidecode==1.3
textdistance==4.2.2
textract==1.6.4
thinc==7.4.5
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
tqdm==4.62.3
troposphere==3.1.1
typer==0.4.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==2.0.1
wordcloud==1.8.1
wsgi-request-logger==0.4.6
xlrd==1.2.0
XlsxWriter==3.0.2
yarg==0.1.9
zappa==0.54.0
zipp==3.6.0
